+++ pbom/io/lzh/compressionchunk.h
@@ -11,7 +11,7 @@ namespace pboman3 {
 
         int flush(QFileDevice* target);
     private:
-        inline static int chunks_ = 8;
+        inline static qint8 chunks_ = 8;
         inline static qint64 minBytesToPack_ = 3;
         inline static qint64 maxChunkSize_ = minBytesToPack_ + 0b1111;
         inline static qint64 maxOffsetToUseWhitespaces_ = CompressionBuffer::defaultSize - maxChunkSize_;
@@ -21,9 +21,9 @@ namespace pboman3 {
         qint8 format_;
         int length_;
 
-        qint64 composeUncompressed(int chunk, QFileDevice* source, CompressionBuffer& dict);
+        qint64 composeUncompressed(qint8 chunk, QFileDevice* source, CompressionBuffer& dict);
 
-        qint64 composeCompressed(int chunk, QFileDevice* source, qint64 chunkSize, CompressionBuffer& dict);
+        qint64 composeCompressed(qint8 chunk, QFileDevice* source, qint64 chunkSize, CompressionBuffer& dict);
 
         qint16 composePointer(qint64 offset, qint64 length);
     };
